+++ Sources/VCore/Views/SwiftUI/Fetch Delegating Completion Image/FetchDelegatingCompletionImageUIModel.swift
@@ -14,7 +14,10 @@ public struct FetchDelegatingCompletionImageUIModel {
     /// Placeholder color.
     public var placeholderColor: Color = .gray.opacity(0.3)
 
-    /// Indicates if image is removed when view disappears. Set to `false`.
+    /// Indicates if `Image` is removed when parameter changes. Set to `true`.
+    public var removesImageOnParameterChange: Bool = true
+
+    /// Indicates if `Image` is removed when component disappears. Set to `false`.
     public var removesImageOnDisappear: Bool = false
 
     // MARK: Initializers